Refactor preview panel with component splitting and risk analysis
- Split preview panel into modular components for better maintainability
- Created preview-panel-components.tsx with extracted reusable components
- Enhanced preview panel with improved structure and performance
- Added comprehensive refactoring risk analysis documentation
- Updated refactoring guide with implementation details and best practices
- Improved code organization and component reusability
- Better separation of concerns in preview panel architecture
