+++ qcut/docs/task/preview-panel-refactoring-guide.md
@@ -3,7 +3,7 @@
 **Date:** 2025-08-07  
 **Branch:** refactor/large-files  
 **Target:** Split `preview-panel.tsx` (1,063 lines) into 2 files  
-**Risk Level:** 🟢 LOW  
+**Risk Level:** 🟡 MEDIUM (with critical fixes applied)  
 **Estimated Time:** 1-2 hours
 
 ## 🎯 Objective
@@ -45,9 +45,12 @@ Lines 888-1063: PreviewToolbar() function
 
 Create new file: `apps/web/src/components/editor/preview-panel-components.tsx`
 
-### Step 2: Set Up Imports for New File
+### ⚠️ CRITICAL WARNING
+**The original imports list was incomplete and would cause compilation failures. Use the CORRECTED imports in Step 2.**
 
-Add these imports to `preview-panel-components.tsx`:
+### Step 2: Set Up Imports and Types for New File
+
+⚠️ **CRITICAL:** Add these imports to `preview-panel-components.tsx`:
 
 ```typescript
 "use client";
@@ -65,14 +68,20 @@ import { useState, useRef, useEffect, useCallback, useMemo } from "react";
 import { usePlaybackStore } from "@/stores/playback-store";
 import { useEditorStore } from "@/stores/editor-store";
 import { useProjectStore } from "@/stores/project-store";
+import { useAspectRatio } from "@/hooks/use-aspect-ratio"; // 🔴 CRITICAL - WAS MISSING
 import { cn } from "@/lib/utils";
 import { formatTimeCode } from "@/lib/time";
 import { EditableTimecode } from "@/components/ui/editable-timecode";
-import { FONT_CLASS_MAP } from "@/lib/font-config";
 import { BackgroundSettings } from "../background-settings";
 import { TimelineElement, TimelineTrack } from "@/types/timeline";
 import type { MediaItem } from "@/stores/media-store-types";
-import { TextElementDragState } from "@/types/editor";
+
+// 🔴 CRITICAL - ADD THIS INTERFACE (used by FullscreenPreview)
+interface ActiveElement {
+  element: TimelineElement;
+  track: TimelineTrack;
+  mediaItem: MediaItem | null;
+}
 ```
 
 ### Step 3: Move Component Functions
@@ -180,21 +189,39 @@ import {
 
 Delete lines 647-1063 from `preview-panel.tsx` (the 3 component functions).
 
-## ✅ Verification Checklist
+## 🚨 Critical Fixes Applied
+
+**Without these fixes, the refactoring WILL BREAK:**
+
+1. **Added missing import:** `useAspectRatio` hook (used by PreviewToolbar)
+2. **Added missing interface:** `ActiveElement` (used by FullscreenPreview)
+3. **Removed unused imports:** `FONT_CLASS_MAP`, `TextElementDragState` (only used in main component)
+
+## ✅ Enhanced Verification Checklist
 
 After completing the refactoring, verify:
 
+**TypeScript Compilation:**
 - [ ] Both files compile without errors
+- [ ] No missing import errors for `useAspectRatio`
+- [ ] No missing type errors for `ActiveElement`
+- [ ] All hook imports resolve correctly
+
+**Functionality Tests:**
 - [ ] Preview panel displays correctly
 - [ ] Fullscreen mode works properly
-- [ ] Preview toolbar controls work
+- [ ] Preview toolbar controls work (aspect ratio dropdown)
 - [ ] Play/pause functionality works
-- [ ] Timeline scrubbing works
-- [ ] Canvas size controls work
-- [ ] Dropdown menus work
-- [ ] All keyboard shortcuts still work
-- [ ] No TypeScript errors
-- [ ] No missing imports
+- [ ] Timeline scrubbing works in both modes
+- [ ] Canvas size controls work (presets dropdown)
+- [ ] Skip forward/backward buttons work
+- [ ] Fullscreen entry/exit works
+
+**Component Interaction Tests:**
+- [ ] PreviewToolbar → FullscreenToolbar transition works
+- [ ] FullscreenPreview renders FullscreenToolbar correctly
+- [ ] All dropdown menus open and function
+- [ ] No console errors or warnings
 
 ## 📁 Final File Structure
 
@@ -221,8 +248,16 @@ export function PreviewPanel() {
 ```typescript
 "use client";
 
-// Required imports for sub-components...
+// All required imports including useAspectRatio...
 
+// ActiveElement interface definition
+interface ActiveElement {
+  element: TimelineElement;
+  track: TimelineTrack;
+  mediaItem: MediaItem | null;
+}
+
+// Components in dependency order:
 export function FullscreenToolbar({ ... }) { ... }
 export function FullscreenPreview({ ... }) { ... }  
 export function PreviewToolbar({ ... }) { ... }
@@ -252,12 +287,20 @@ After successfully completing this refactoring:
 2. Consider applying same pattern to `export-dialog.tsx` (next largest file)
 3. Document the pattern for future large file refactoring
 
-## ⚠️ Important Notes
+## ⚠️ Critical Success Factors
 
 - **Don't change any logic** - Only move code, don't modify functionality
 - **Keep all prop interfaces identical** - Don't change component APIs
-- **Preserve all imports** - Ensure all dependencies are available in both files
+- **MUST include critical fixes** - Missing imports/types will cause compilation failures
 - **Test thoroughly** - Preview panel is critical editor functionality
 - **Export all moved functions** - Make sure they're properly exported from new file
+- **Component order matters** - Follow dependency order in new file
 
-This refactoring is safe because it only reorganizes code without changing any business logic or component interfaces.
+## 🔴 Common Pitfalls to Avoid
+
+1. **Forgetting `useAspectRatio` import** - PreviewToolbar will break
+2. **Missing `ActiveElement` interface** - FullscreenPreview will fail to compile
+3. **Wrong component order** - May cause hoisting issues
+4. **Leaving unused imports** - Keep imports clean and relevant
+
+This refactoring is safe **ONLY** with the critical fixes applied. The original plan had missing dependencies that would cause immediate compilation failures.
