+++ qcut/docs/task/preview-panel-refactoring-risk-analysis.md
@@ -1,0 +1,169 @@
+# Preview Panel Refactoring Risk Analysis
+
+**Date:** 2025-08-07  
+**Branch:** refactor/large-files  
+**Analysis:** Deep dive into potential breaking points
+
+## 🔍 Ultra-Deep Risk Analysis
+
+After thorough examination of `preview-panel.tsx`, here are the **potential breaking points** and **critical fixes** needed:
+
+## ❌ **CRITICAL ISSUES FOUND**
+
+### 1. **MISSING IMPORT: `useAspectRatio`**
+**Risk Level:** 🔴 **HIGH - WILL BREAK**
+
+**Problem:** The `PreviewToolbar` component uses `useAspectRatio` hook (line 914), but it's not in my original import list.
+
+**Fix Required:**
+```typescript
+// ADD this import to preview-panel-components.tsx
+import { useAspectRatio } from "@/hooks/use-aspect-ratio";
+```
+
+### 2. **MISSING TYPE: `ActiveElement`**
+**Risk Level:** 🔴 **HIGH - WILL BREAK**
+
+**Problem:** The `FullscreenPreview` component uses `ActiveElement[]` type (line 831), but `ActiveElement` interface is defined in the main file and won't be available in the new file.
+
+**Fix Required:**
+```typescript
+// ADD this interface to preview-panel-components.tsx  
+interface ActiveElement {
+  element: TimelineElement;
+  track: TimelineTrack;
+  mediaItem: MediaItem | null;
+}
+```
+
+## ⚠️ **MEDIUM RISK ISSUES**
+
+### 3. **Internal Component Dependencies**
+**Risk Level:** 🟡 **MEDIUM - NEEDS ATTENTION**
+
+**Discovered Dependencies:**
+- `FullscreenPreview` calls `FullscreenToolbar` (line 875)
+- `PreviewToolbar` calls `FullscreenToolbar` (line 939)
+
+**Why this is OK:** Since all components move to the same file, internal dependencies remain intact.
+
+**Verification needed:** Ensure component order doesn't create hoisting issues.
+
+## ✅ **VERIFIED SAFE ASPECTS**
+
+### 4. **No Module-Level Constants Used** ✅
+- `debugLogger` - only used in main component (line 524)
+- `FONT_CLASS_MAP` - only used in main component (line 382)
+- ✅ **Safe:** Functions to be moved don't use these constants
+
+### 5. **No Closure Dependencies** ✅
+- All three functions are defined at module level (not nested)
+- All data passed via props, no closure variable access
+- ✅ **Safe:** No scope issues
+
+## 🛠️ **CORRECTED REFACTORING PLAN**
+
+### Updated Import List for `preview-panel-components.tsx`:
+
+```typescript
+"use client";
+
+import { Button } from "@/components/ui/button";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+  DropdownMenuSeparator,
+} from "@/components/ui/dropdown-menu";
+import { Play, Pause, Expand, SkipBack, SkipForward } from "lucide-react";
+import { useState, useRef, useEffect, useCallback, useMemo } from "react";
+import { usePlaybackStore } from "@/stores/playback-store";
+import { useEditorStore } from "@/stores/editor-store";
+import { useProjectStore } from "@/stores/project-store";
+import { useAspectRatio } from "@/hooks/use-aspect-ratio"; // 🔴 CRITICAL - WAS MISSING
+import { cn } from "@/lib/utils";
+import { formatTimeCode } from "@/lib/time";
+import { EditableTimecode } from "@/components/ui/editable-timecode";
+import { BackgroundSettings } from "../background-settings";
+import { TimelineElement, TimelineTrack } from "@/types/timeline";
+import type { MediaItem } from "@/stores/media-store-types";
+
+// 🔴 CRITICAL - ADD THIS INTERFACE
+interface ActiveElement {
+  element: TimelineElement;
+  track: TimelineTrack;
+  mediaItem: MediaItem | null;
+}
+```
+
+### Component Order in New File:
+```typescript
+// 1. FullscreenToolbar first (no dependencies)
+export function FullscreenToolbar({ ... }) { ... }
+
+// 2. FullscreenPreview second (depends on FullscreenToolbar)  
+export function FullscreenPreview({ ... }) { ... }
+
+// 3. PreviewToolbar last (depends on FullscreenToolbar)
+export function PreviewToolbar({ ... }) { ... }
+```
+
+## 🧪 **ENHANCED TESTING CHECKLIST**
+
+**Critical areas to test after refactoring:**
+
+### Functionality Tests:
+- [ ] **FullscreenToolbar:** Timeline scrubbing, play/pause, skip buttons
+- [ ] **FullscreenPreview:** Full-screen mode entry/exit, rendering
+- [ ] **PreviewToolbar:** Canvas size presets, aspect ratio changes
+- [ ] **Component interactions:** PreviewToolbar → FullscreenToolbar transition
+- [ ] **Component interactions:** FullscreenPreview → FullscreenToolbar rendering
+
+### TypeScript Tests:
+- [ ] No TypeScript errors in both files
+- [ ] `ActiveElement` type properly resolved
+- [ ] All hook imports working correctly
+
+### Runtime Tests:
+- [ ] All dropdown menus functional
+- [ ] Timeline interactions smooth
+- [ ] No console errors
+- [ ] Component state transitions work
+
+## 🚨 **RED FLAGS TO WATCH**
+
+1. **TypeScript errors about missing `ActiveElement`** - Check interface is copied
+2. **Import errors for `useAspectRatio`** - Check import is added
+3. **Component render errors** - Check component order and dependencies
+4. **Hook errors** - Check all hook imports are present
+
+## ✅ **UPDATED RISK ASSESSMENT**
+
+**Original Risk:** 🟢 LOW  
+**After Analysis:** 🟡 **MEDIUM** 
+
+**Why medium risk:** Missing critical imports and types that will cause immediate compilation failures if not addressed.
+
+**With fixes applied:** 🟢 **LOW** - Plan is safe with corrections
+
+## 📝 **EXECUTION PLAN**
+
+1. ✅ Apply the corrected import list
+2. ✅ Add the `ActiveElement` interface to new file  
+3. ✅ Ensure proper component ordering
+4. ✅ Move all three components
+5. ✅ Update main file imports
+6. ✅ Test thoroughly with enhanced checklist
+
+## 🎯 **CONCLUSION**
+
+**The refactoring plan is SAFE** if you:
+1. ✅ Add missing `useAspectRatio` import
+2. ✅ Include `ActiveElement` interface definition
+3. ✅ Follow the corrected import list
+
+**Without these fixes:** 🔴 **WILL BREAK** - TypeScript compilation errors  
+**With fixes:** 🟢 **SAFE** - Clean refactoring with no runtime issues
+
+The original plan had critical oversights that would cause immediate failures. With the corrections above, the refactoring becomes safe and straightforward.
